requirements.txt
# Core
langchain>=0.2.0
langgraph>=0.2.6
pydantic>=2.7.1
tenacity>=8.2.3
httpx>=0.27.0
pyyaml>=6.0.1
faiss-cpu>=1.8.0.post1

# Optional / Mock friendly
openai>=1.40.0
slack_sdk>=3.31.0
kubernetes>=30.1.0
fastapi>=0.111.0
uvicorn>=0.30.1
pytest>=8.2.0
loguru>=0.7.2
